Restructured bld_core modules.
