+++ crates/bld_core/src/context/mod.rs
@@ -18,7 +18,7 @@ struct ContextReceiver {
 }
 
 impl ContextReceiver {
-    pub fn containers_atom(
+    pub fn containers(
         pool: Arc<Pool<ConnectionManager<SqliteConnection>>>,
         run_id: &str,
     ) -> Self {
@@ -121,7 +121,7 @@ pub struct ContextSender {
 impl ContextSender {
     pub fn new(pool: Arc<Pool<ConnectionManager<SqliteConnection>>>, run_id: &str) -> Self {
         let (tx, rx) = channel(4096);
-        let context = ContextReceiver::containers_atom(pool, run_id);
+        let context = ContextReceiver::containers(pool, run_id);
 
         spawn(async move {
             if let Err(e) = context.receive(rx).await {
