--- crates/bld_core/src/scanner/base.rs
@@ -1,3 +1,0 @@
-pub trait Scanner {
-    fn fetch(&mut self) -> Vec<String>;
-}
