+++ crates/bld_server/src/sockets/monit.rs
@@ -7,7 +7,7 @@ use actix_web_actors::ws;
 use anyhow::{anyhow, Result};
 use bld_config::BldConfig;
 use bld_core::database::pipeline_runs::{self, PR_STATE_FAULTED, PR_STATE_FINISHED};
-use bld_core::scanner::{FileScanner, Scanner};
+use bld_core::scanner::FileScanner;
 use bld_sock::messages::MonitInfo;
 use diesel::r2d2::{ConnectionManager, Pool};
 use diesel::sqlite::SqliteConnection;
@@ -35,7 +35,7 @@ impl MonitorPipelineSocket {
 
     fn scan(act: &mut Self, ctx: &mut <Self as Actor>::Context) {
         if let Some(scanner) = act.scanner.as_mut() {
-            let content = scanner.fetch();
+            let content = scanner.scan();
             for line in content.iter() {
                 ctx.text(line.to_string());
             }
